+++ src/vs/platform/configuration/test/common/testConfigurationService.ts
@@ -9,7 +9,7 @@ import { TernarySearchTree } from 'vs/base/common/map';
 import URI from 'vs/base/common/uri';
 import { TPromise } from 'vs/base/common/winjs.base';
 import { EventEmitter } from 'vs/base/common/eventEmitter';
-import { getConfigurationKeys, IConfigurationOverrides, IConfigurationService, getConfigurationValue } from 'vs/platform/configuration/common/configuration';
+import { getConfigurationKeys, IConfigurationOverrides, IConfigurationService, getConfigurationValue, isConfigurationOverrides } from 'vs/platform/configuration/common/configuration';
 
 export class TestConfigurationService extends EventEmitter implements IConfigurationService {
 	public _serviceBrand: any;
@@ -22,7 +22,8 @@ export class TestConfigurationService extends EventEmitter implements IConfigura
 		return TPromise.as(this.getConfiguration());
 	}
 
-	public getConfiguration<C>(section?: any, overrides?: any): C {
+	public getConfiguration<C>(arg1?: any, arg2?: any): C {
+		const overrides = isConfigurationOverrides(arg1) ? arg1 : isConfigurationOverrides(arg2) ? arg2 : void 0;
 		if (overrides && overrides.resource) {
 			const configForResource = this.configurationByRoot.findSubstr(overrides.resource.fsPath);
 			return configForResource || this.configuration;
